fix(security): single isPrivateAddress SSOT — the SSRF guard was forked
Sweep finding #3 (HIGH): the private-address check existed as two
independently-maintained copies — lib/security/ssrfGuard.ts and
services/cat/website-analysis.ts (the user-supplied-URL surface, i.e. the
highest-risk SSRF path). A range added to one and not the other would
have opened a silent hole; they had already drifted (the security module
was MISSING IPv6 multicast ff00::/8, which the Cat copy blocked).

- New src/lib/security/private-address.ts: the single, dependency-free
  implementation (strictest union of both copies: brackets handled,
  malformed input fail-closed, ff00::/8 + site-local covered). No
  node:net import, so client-bundle-safe modules can use it too.
- ssrfGuard.ts and website-analysis.ts both import/re-export it; their
  local copies are deleted.

All 102 ssrfGuard + website-analysis tests green, tsc + eslint clean.
