+++ src/lib/security/private-address.ts
@@ -1,0 +1,61 @@
+/**
+ * isPrivateAddress — THE single implementation of "is this IP something a
+ * server-side fetch must never touch" (loopback, RFC1918, link-local/cloud
+ * metadata, CGNAT, ULA, multicast, reserved, v4-mapped v6).
+ *
+ * Deliberately dependency-free (no node:net) so it can be imported from any
+ * module, including ones that must stay client-bundle-safe. Fail-closed: any
+ * string that isn't a well-formed address is treated as private.
+ *
+ * History: this logic existed as two independently-maintained copies
+ * (lib/security/ssrfGuard.ts and services/cat/website-analysis.ts) — a range
+ * added to one and not the other would have opened a silent SSRF hole. Both
+ * now import from here.
+ */
+export function isPrivateAddress(ip: string): boolean {
+  const addr = ip.trim().toLowerCase();
+
+  // IPv6 (including v4-mapped ::ffff:a.b.c.d)
+  if (addr.includes(':')) {
+    const unbracketed = addr.replace(/^\[|\]$/g, '');
+    const mapped = unbracketed.match(/^::ffff:(\d+\.\d+\.\d+\.\d+)$/);
+    if (mapped) {
+      return isPrivateAddress(mapped[1]);
+    }
+    if (unbracketed === '::' || unbracketed === '::1') {
+      return true; // unspecified / loopback
+    }
+    if (/^fe[89ab]/.test(unbracketed)) {
+      return true; // link-local fe80::/10
+    }
+    if (/^f[cd]/.test(unbracketed)) {
+      return true; // unique-local fc00::/7
+    }
+    if (/^fe[cdef]/.test(unbracketed)) {
+      return true; // deprecated site-local fec0::/10
+    }
+    if (/^ff/.test(unbracketed)) {
+      return true; // multicast ff00::/8
+    }
+    return false;
+  }
+
+  // IPv4
+  const octets = addr.split('.').map(Number);
+  if (octets.length !== 4 || octets.some(o => Number.isNaN(o) || o < 0 || o > 255)) {
+    return true; // not a well-formed IPv4 — refuse rather than guess
+  }
+  const [a, b] = octets;
+  return (
+    a === 0 || // 0.0.0.0/8 "this network"
+    a === 10 || // 10.0.0.0/8 private
+    a === 127 || // 127.0.0.0/8 loopback
+    (a === 100 && b >= 64 && b <= 127) || // 100.64.0.0/10 CGNAT
+    (a === 169 && b === 254) || // 169.254.0.0/16 link-local (cloud metadata!)
+    (a === 172 && b >= 16 && b <= 31) || // 172.16.0.0/12 private
+    (a === 192 && b === 0) || // 192.0.0.0/24 + 192.0.2.0/24 special/doc
+    (a === 192 && b === 168) || // 192.168.0.0/16 private
+    (a === 198 && (b === 18 || b === 19)) || // 198.18.0.0/15 benchmarking
+    a >= 224 // multicast 224/4 + reserved 240/4 + broadcast
+  );
+}
